requirements.txt
pytubefix>=2.5.0
moviepy>=1.0.3
Pillow>=10.0.0
requests>=2.31.0